requirements.txt
aiofiles==24.1.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.112.2
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.8
kombu==5.4.0
numpy==2.1.0
opencv-python==4.10.0.84
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
redis==5.0.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
